requirements.txt
Werkzeug
flask
redis
numpy 
benford_py== 0.2.7
scipy
pandas
